Email sender/requirements.txt
python-dotenv
smtplib
ssl
email.mime.text
email.mime.multipart